// src/generated/dokploy-schemas.ts
export const procedureSchemas: Record<string, ProcedureSchema> = {
// ...
  "application.one": {
    "method": "GET",
    "path": "/application.one",
    "tag": "application",
    "inputKind": "query",
    "inputSchema": {
      "type": "object",
      "properties": {
        "applicationId": {
          "type": "string",
          "minLength": 1
        }
      },
      "required": [
        "applicationId"
      ]
    },
    "outputSchema": {
      "type": "object",
      "properties": {},
      "additionalProperties": false
    }
  },
// ...
  "application.saveGithubProvider": {
    "method": "POST",
    "path": "/application.saveGithubProvider",
    "tag": "application",
    "inputKind": "body",
    "inputSchema": {
      "type": "object",
      "properties": {
// ...
        "githubId": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        },
// ...
      },
      "required": [
        "applicationId",
        "repository",
        "branch",
        "owner",
        "buildPath",
        "githubId",
        "triggerType"
      ]
    },
    "outputSchema": {
      "type": "object",
      "properties": {},
      "additionalProperties": false
    }
  },
  "application.saveGitlabProvider": {
    "method": "POST",
    "path": "/application.saveGitlabProvider",
    "tag": "application",
    "inputKind": "body",
    "inputSchema": {
      "type": "object",
      "properties": {
// ...
        "gitlabId": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        },
// ...
      },
      "required": [
        "applicationId",
        "gitlabBranch",
        "gitlabBuildPath",
        "gitlabOwner",
        "gitlabRepository",
        "gitlabId",
        "gitlabProjectId",
        "gitlabPathNamespace"
      ]
    },
    "outputSchema": {
      "type": "object",
      "properties": {},
      "additionalProperties": false
    }
  },
  "application.saveBitbucketProvider": {
    "method": "POST",
    "path": "/application.saveBitbucketProvider",
    "tag": "application",
    "inputKind": "body",
    "inputSchema": {
      "type": "object",
      "properties": {
// ...
        "bitbucketId": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        },
        "applicationId": {
          "type": "string"
        },
        "enableSubmodules": {
          "type": "boolean"
        },
        "watchPaths": {
          "anyOf": [
            {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            {
              "type": "null"
            }
          ]
        }
      },
      "required": [
        "bitbucketBranch",
        "bitbucketBuildPath",
        "bitbucketOwner",
        "bitbucketRepository",
        "bitbucketRepositorySlug",
        "bitbucketId",
        "applicationId"
      ]
    },
    "outputSchema": {
      "type": "object",
      "properties": {},
      "additionalProperties": false
    }
  },
  "application.saveGiteaProvider": {
    "method": "POST",
    "path": "/application.saveGiteaProvider",
    "tag": "application",
    "inputKind": "body",
    "inputSchema": {
      "type": "object",
      "properties": {
// ...
        "giteaId": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "null"
            }
          ]
        },
        "enableSubmodules": {
          "type": "boolean"
        },
        "watchPaths": {
          "anyOf": [
            {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            {
              "type": "null"
            }
          ]
        }
      },
      "required": [
        "applicationId",
        "giteaBranch",
        "giteaBuildPath",
        "giteaOwner",
        "giteaRepository",
        "giteaId"
      ]
    },
    "outputSchema": {
      "type": "object",
      "properties": {},
      "additionalProperties": false
    }
  },
// ...
  "certificates.create": {
    "method": "POST",
    "path": "/certificates.create",
    "tag": "certificates",
    "inputKind": "body",
    "inputSchema": {
      "type": "object",
      "properties": {
        "certificateId": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "minLength": 1
        },
        "certificateData": {
          "type": "string",
          "minLength": 1
        },
        "privateKey": {
          "type": "string",
          "minLength": 1
        },
// ...
      },
      "required": [
        "name",
        "certificateData",
        "privateKey",
        "organizationId"
      ]
    },
    "outputSchema": {
      "type": "object",
      "properties": {},
      "additionalProperties": false
    }
  },
// ...
}